requirements.txt
opencv-python
tensorflow
matplotlib
numpy
TK
sklearn
pandas
Pillow